# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Fitness RPG API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training/exercises:
    get:
      operationId: getExercises
      tags: [training]
      summary: List exercises
      parameters:
        - name: equipmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: muscleGroup
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /training/templates:
    get:
      operationId: getWorkoutTemplates
      tags: [training]
      summary: List workout templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutTemplate"
    post:
      operationId: createWorkoutTemplate
      tags: [training]
      summary: Create workout template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutTemplate"
  /training/templates/{id}:
    get:
      operationId: getWorkoutTemplate
      tags: [training]
      summary: Get a workout template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutTemplate"
    patch:
      operationId: updateWorkoutTemplate
      tags: [training]
      summary: Update workout template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutTemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutTemplate"
    delete:
      operationId: deleteWorkoutTemplate
      tags: [training]
      summary: Delete workout template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /training/sessions:
    get:
      operationId: getWorkoutSessions
      tags: [training]
      summary: List workout sessions
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutSession"
    post:
      operationId: createWorkoutSession
      tags: [training]
      summary: Start a workout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSession"
  /training/sessions/{id}:
    get:
      operationId: getWorkoutSession
      tags: [training]
      summary: Get a workout session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSession"
    patch:
      operationId: updateWorkoutSession
      tags: [training]
      summary: Update/complete a workout session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionUpdate"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSession"
  /training/sessions/{id}/sets:
    post:
      operationId: logSet
      tags: [training]
      summary: Log a set in a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSetInput"
      responses:
        "201":
          description: Logged set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSet"
  /training/sessions/{id}/sets/{setId}:
    patch:
      operationId: updateSet
      tags: [training]
      summary: Update a logged set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSetUpdate"
      responses:
        "200":
          description: Updated set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSet"
    delete:
      operationId: deleteSet
      tags: [training]
      summary: Delete a logged set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Exercise:
      type: object
      required: [id, name, muscleGroup, category]
      properties:
        id:
          type: integer
        name:
          type: string
        muscleGroup:
          type: string
        category:
          type: string
          enum: [barbell, dumbbell, machine, bodyweight, cable, cardio, martial_arts]
        instructions:
          type: ["string", "null"]
        equipmentIds:
          type: array
          items:
            type: integer
    WorkoutTemplate:
      type: object
      required: [id, name, category, exercises, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [strength, conditioning, striking, grappling, recovery, mixed]
        description:
          type: ["string", "null"]
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/TemplateExercise"
        estimatedDuration:
          type: ["integer", "null"]
        xpReward:
          type: integer
        createdAt:
          type: string
    TemplateExercise:
      type: object
      required: [exerciseId, exerciseName, sets, reps, order]
      properties:
        exerciseId:
          type: integer
        exerciseName:
          type: string
        sets:
          type: integer
        reps:
          type: string
        restSeconds:
          type: ["integer", "null"]
        order:
          type: integer
        notes:
          type: ["string", "null"]
    WorkoutTemplateInput:
      type: object
      required: [name, category, exercises]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [strength, conditioning, striking, grappling, recovery, mixed]
        description:
          type: ["string", "null"]
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/TemplateExercise"
        estimatedDuration:
          type: ["integer", "null"]
    WorkoutTemplateUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [strength, conditioning, striking, grappling, recovery, mixed]
        description:
          type: ["string", "null"]
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/TemplateExercise"
        estimatedDuration:
          type: ["integer", "null"]
    WorkoutSession:
      type: object
      required: [id, name, status, sets, startedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        templateId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, completed, abandoned]
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
        xpEarned:
          type: ["integer", "null"]
        goldEarned:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
    WorkoutSessionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        templateId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    WorkoutSessionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, abandoned]
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
    WorkoutSet:
      type: object
      required: [id, sessionId, exerciseId, exerciseName, setNumber, reps, weight, weightUnit, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        setNumber:
          type: integer
        reps:
          type: integer
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        rpe:
          type: ["number", "null"]
        isPr:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkoutSetInput:
      type: object
      required: [exerciseId, setNumber, reps, weight, weightUnit]
      properties:
        exerciseId:
          type: integer
        setNumber:
          type: integer
        reps:
          type: integer
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        rpe:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    WorkoutSetUpdate:
      type: object
      properties:
        reps:
          type: integer
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        rpe:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...