# lib/api-zod/src/generated/types/workoutSetUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Fitness RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutSetUpdateWeightUnit } from './workoutSetUpdateWeightUnit';

export interface WorkoutSetUpdate {
  reps?: number;
  weight?: number;
  weightUnit?: WorkoutSetUpdateWeightUnit;
  /** @nullable */
  rpe?: number | null;
  /** @nullable */
  notes?: string | null;
}